Update BuildView.vue
Unpredictable painting bug was a result of not awaiting properly for this click callback. Somehow, somewhere, the pixel was being reset to the original location without any notice. Maybe the store is being accessed as things are being removed? Not worth looking into, but must remember async for strange bugs.