Create 0003_voxel_updated_at.py
note: don't forget to ssh into the production vm to migrate after redeployment. The timing on this always risks breaking the production site for the amount of time it takes for the new vm to spin up, have the migrations applied, then restart. I think one solution for this is having multiple deployment slots, but the cost for that is maybe not worth it at this stage.