Update BuildView.vue
Update builder grid to be 16x16 instead of 8x8.

This makes building navigation easier, although the placement grid buttons are smaller. I think this is a good balance; we'll try it for a while. Fine on desktop in dev and will see how it feels on the phone after deployment.