Update BuildView.vue
Fix flipped 2D builder display.

I think this is the simplest, most straightforward fix to address the flipped/twisted issue between 3D world and 2D builder interface. Repositioning the builder will now display the 2D plane slices in the same orientation as they are in the 3D world, instead of flipped.

Transforming the interface instead of the logic prevents causing more twists that would need to be rectified between the systems that are working in sync.